+++ Resources/FormBody/CommonGroupTaskFormData.cs
@@ -2,8 +2,8 @@
 
 public class CommonGroupTaskFormData
 {
-  public float Start { get; set; }
-  public float Duration { get; set; }
+  public decimal Start { get; set; }
+  public decimal Duration { get; set; }
   public string Name { set; get; } = null!;
   public string Id { set; get; } = null!;
   public string Type { set; get; } = null!;
@@ -16,5 +16,5 @@ public class CommonGroupTaskFormData
   public string? Note { get; set; } = null!;
   public ICollection<StepworkResource>? Stepworks { get; set; }
   public ICollection<PredecessorResource>? Predecessors { get; set; }
-  public float End { get; set; }
+  public decimal End { get; set; }
 }
